refactor: extract net worth metric
